feat(mandev-react): 调整项目结构(changelog-needed)
ISSUES CLOSED: none
