+++ src/test/App.test.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import './setupTests'
+import { render, screen } from '@testing-library/react'
+import App from '../App'
+
+test('renders learn react link', () => {
+	render(<App />)
+	const linkElement = screen.getByText(/learn react/i)
+	expect(linkElement).toBeInTheDocument()
+})
